refactor(best_sellers): restructure best sellers screen and improve responsiveness
