+++ lib/main.dart
@@ -43,7 +43,7 @@ class MyApp extends StatelessWidget {
         title: 'Flutter Demo',
         routes: AppRoutes.routes,
         theme: AppTheme.appTheme,
-        initialRoute: AppRoutes.mainLayoutRoute,
+        initialRoute: AppRoutes.bestSellerRoute,
       ),
     );
   }
